cleanup commitlog integration tests
